+++ AoDDownloader/config.py
@@ -2,7 +2,10 @@ import json
 import os
 from pathlib import Path
 
+import click
+
 from .quality import Quality
+
 
 class Config(object):
     APPKEY = "AoDDownloader"
@@ -26,7 +29,7 @@ class Config(object):
                 "username": self.username,
                 "german": self.german,
                 "japanese": self.japanese,
-                "quality": self.quality.name
+                "quality": self.quality.name if self.quality else ""
             }
             json.dump(config_dict, config_file)
 
@@ -36,3 +39,15 @@ class Config(object):
         if not os.path.exists(app_path):
             os.mkdir(app_path)
         return app_path + "/config.json"
+
+    def setQuality(self):
+        self.quality = Quality[
+            click.prompt("Select the max wanted quality.", type=click.Choice([e.name for e in Quality]),
+                         show_choices=True)]
+        self.write()
+
+    def setLanguages(self):
+        self.japanese = click.confirm("Try downloading japanese audio with subtitles?")
+        self.german = click.confirm("Try downloading german audio?")
+        self.write()
+        return self.japanese, self.german