+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 - use the best quality available if max quality is not met
 - get requirements from requirements.txt
 - ensure Python version 3.8 or higher
+- add config changing options
+- refactor language/quality selection
 
 ## 0.5.0
 - use click for progressbar